version 2.3.8

+++ entrypoint.sh
@@ -8,9 +8,22 @@ export EXAKAT_INCLUDE_DIRS=$INPUT_INCLUDE_DIRS
 export EXAKAT_FILE_EXTENSIONS=$INPUT_FILE_EXTENSIONS
 export EXAKAT_PROJECT_REPORTS=$INPUT_PROJECT_REPORTS
 
-ln -s /github/workspace projects/review/code
+if [ ! -s projects/review/code ] 
+then
+    ln -s /github/workspace projects/review/code
+fi
 
-exakat project -p review -v "$@"
+exakat project -p review "$@"
+
+if [ -f projects/review/exakat.sarif ] 
+then
+    mv projects/review/exakat.sarif /github/workspace/exakat.sarif
+fi
+
+if [ -d projects/review/diplomat ] 
+then
+    mv projects/review/diplomat /github/workspace/diplomat
+fi
 
 # check 
 
@@ -19,7 +32,7 @@ then
     # get the results from Text format, to count them
     exakat report -p review --format Text > exit.txt
     if [ -s exit.txt ] 
-    then
+    then    
         exit 1
     else
         exit 0
